+++ Changelog.md
@@ -2,14 +2,26 @@
 
 ## [Unreleased]
 
+### [3.11.1] - 2026-06-18
+### Added
+- Show tile for zkp
+- show zkp error message
+- change the zkp label to normal user readable text
+- added toggle button for widget zkp configure
+- Added toggel to disable widget login
+
+### Added
+- Added toggel to disable widget login
+### [3.11.0] - 2026-06-18
+
+### Added
+- Added toggel to disable widget login
 ### [3.11.0] - 2026-06-18
 
 ### Added
 - Added toggel to disable widget login
 
-
 ### [3.10.5] - 2026-06-11
-
 
 ### Fixed
 - Fixed busines icon